Issue #367 - Move FromFuture to core and leave only typeclass instances in the sub-projects
