+++ cats-effect3/src/main/scala-3/effectie/cats/fromFuture.scala
@@ -2,6 +2,8 @@ package effectie.cats
 
 import cats.Id
 import cats.effect.IO
+import effectie.core.FromFuture
+import effectie.core.FromFuture.FromFutureToIdTimeout
 
 import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, Future}
@@ -9,23 +11,11 @@ import scala.concurrent.{Await, Future}
 /** @author Kevin Lee
   * @since 2020-09-22
   */
-trait FromFuture[F[_]] {
-  def toEffect[A](future: => Future[A]): F[A]
-}
-
-object FromFuture {
-
-  def apply[F[_]: FromFuture]: FromFuture[F] = summon[FromFuture[F]]
+object fromFuture {
 
   given fromFutureToIo: FromFuture[IO] with {
     override def toEffect[A](future: => Future[A]): IO[A] = IO.fromFuture[A](IO(future))
   }
-
-  given fromFutureToFuture: FromFuture[Future] with {
-    override def toEffect[A](future: => Future[A]): Future[A] = future
-  }
-
-  final case class FromFutureToIdTimeout(fromFutureToIdTimeout: Duration)
 
   given fromFutureToId(using timeout: FromFutureToIdTimeout): FromFuture[Id] with {
     override def toEffect[A](future: => Future[A]): Id[A] =
