+++ cats-effect3/src/test/scala-2/effectie/cats/fromFutureSpec.scala
@@ -4,12 +4,13 @@ import cats.Id
 import cats.effect._
 import cats.effect.unsafe.IORuntime
 import effectie.cats.compat.CatsEffectIoCompatForFuture
+import effectie.cats.fromFuture._
+import effectie.core.FromFuture
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
 import hedgehog.runner._
 
-import java.util.concurrent.ExecutorService
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -21,7 +22,6 @@ object FromFutureSpec extends Properties {
 
   override def tests: List[Test] = List(
     property("test FromFuture[IO].toEffect", IoSpec.testToEffect),
-    property("test FromFuture[Future].toEffect", FutureSpec.testToEffect),
     property("test FromFuture[Id].toEffect", IdSpec.testToEffect)
   )
 
@@ -39,23 +39,6 @@ object FromFutureSpec extends Properties {
       ConcurrentSupport.runAndShutdown(compat.es, waitFor300Millis) {
         lazy val fa = Future(a)
         val actual  = FromFuture[IO].toEffect(fa).unsafeRunSync()
-
-        actual ==== a
-      }
-    }
-  }
-
-  object FutureSpec {
-    def testToEffect: Property = for {
-      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
-    } yield {
-      implicit val es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
-      implicit val ec: ExecutionContext =
-        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
-
-      ConcurrentSupport.runAndShutdown(es, waitFor300Millis) {
-        lazy val fa = Future(a)
-        val actual  = ConcurrentSupport.futureToValueAndTerminate(es, waitFor300Millis)(FromFuture[Future].toEffect(fa))
 
         actual ==== a
       }
