+++ cats-effect3/src/test/scala-3/effectie/cats/fromFutureSpec.scala
@@ -4,6 +4,8 @@ import cats.Id
 import cats.effect.*
 import cats.effect.unsafe.IORuntime
 import effectie.cats.compat.CatsEffectIoCompatForFuture
+import effectie.cats.fromFuture.given
+import effectie.core.FromFuture
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog.*
@@ -16,13 +18,12 @@ import scala.concurrent.{ExecutionContext, Future}
 /** @author Kevin Lee
   * @since 2020-09-22
   */
-object FromFutureSpec extends Properties {
+object fromFutureSpec extends Properties {
 
   private given errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
   override def tests: List[Test] = List(
     property("test FromFuture[IO].toEffect", IoSpec.testToEffect),
-    property("test FromFuture[Future].toEffect", FutureSpec.testToEffect),
     property("test FromFuture[Id].toEffect", IdSpec.testToEffect)
   )
 
@@ -44,28 +45,13 @@ object FromFutureSpec extends Properties {
     }
   }
 
-  object FutureSpec {
-    def testToEffect: Property = for {
-      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
-    } yield {
-      given es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
-      given ec: ExecutionContext = ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
-
-      ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
-        lazy val fa = Future(a)
-        val actual  = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(FromFuture[Future].toEffect(fa))
-
-        actual ==== a
-      }
-    }
-  }
-
   object IdSpec {
     def testToEffect: Property = for {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
     } yield {
       val es                     = ConcurrentSupport.newExecutorService(2)
-      given ec: ExecutionContext = ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
 
       ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
         given timeout: FromFuture.FromFutureToIdTimeout =
