+++ core/src/test/scala-3/effectie/core/FromFutureSpec.scala
@@ -1,0 +1,42 @@
+package effectie.core
+
+import extras.concurrent.testing.ConcurrentSupport
+import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
+import hedgehog.*
+import hedgehog.runner.*
+
+import java.util.concurrent.ExecutorService
+import scala.concurrent.duration.*
+import scala.concurrent.{ExecutionContext, Future}
+
+/** @author Kevin Lee
+  * @since 2020-09-22
+  */
+object FromFutureSpec extends Properties {
+  implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
+
+  private val waitFor300Millis = WaitFor(300.milliseconds)
+
+  override def tests: List[Test] = List(
+    property("test FromFuture[Future].toEffect", FutureSpec.testToEffect),
+  )
+
+  object FutureSpec {
+    def testToEffect: Property = for {
+      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
+    } yield {
+      given es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
+      given ec: ExecutionContext =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+
+      ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
+        lazy val fa = Future(a)
+        val actual  =
+          ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(FromFuture[Future].toEffect(fa))
+
+        actual ==== a
+      }
+    }
+  }
+
+}
