+++ effectie-monix/src/test/scala/effectie/monix/fromFutureSpec.scala
@@ -2,6 +2,8 @@ package effectie.monix
 
 import cats.Id
 import cats.effect.{ContextShift, IO}
+import effectie.core.FromFuture
+import effectie.monix.fromFuture._
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
@@ -9,14 +11,13 @@ import hedgehog.runner._
 import monix.eval.Task
 import monix.execution.Scheduler
 
-import java.util.concurrent.ExecutorService
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
 
 /** @author Kevin Lee
   * @since 2020-09-22
   */
-object FromFutureSpec extends Properties {
+object fromFutureSpec extends Properties {
   private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
   private val waitFor300Millis = WaitFor(300.milliseconds)
@@ -24,7 +25,6 @@ object FromFutureSpec extends Properties {
   override def tests: List[Test] = List(
     property("test FromFuture[Task].toEffect", TaskSpec.testToEffect),
     property("test FromFuture[IO].toEffect", IoSpec.testToEffect),
-    property("test FromFuture[Future].toEffect", FutureSpec.testToEffect),
     property("test FromFuture[Id].toEffect", IdSpec.testToEffect)
   )
 
@@ -58,23 +58,6 @@ object FromFutureSpec extends Properties {
       ConcurrentSupport.runAndShutdown(es, waitFor300Millis) {
         lazy val fa = Future(a)
         val actual  = FromFuture[IO].toEffect(fa).unsafeRunSync()
-
-        actual ==== a
-      }
-    }
-  }
-
-  object FutureSpec {
-    def testToEffect: Property = for {
-      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
-    } yield {
-      implicit val es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
-      implicit val ec: ExecutionContext =
-        ConcurrentSupport.newExecutionContext(es, ErrorLogger.printlnExecutionContextErrorLogger)
-
-      ConcurrentSupport.runAndShutdown(es, waitFor300Millis) {
-        lazy val fa = Future(a)
-        val actual  = ConcurrentSupport.futureToValueAndTerminate(es, waitFor300Millis)(FromFuture[Future].toEffect(fa))
 
         actual ==== a
       }
